+++ client-sdk/src/ring_reader.rs
@@ -507,7 +507,7 @@ fn bucket_entry_to_slice(
 ) -> Result<&[u8], BucketTooShort> {
     let index = usize::try_from(entry.index()).unwrap();
     let size = usize::try_from(entry.size()).unwrap();
-    let bucket = size_to_bucket(entry.size());
+    let bucket = usize::from(size_to_bucket(entry.size()));
 
     let size_class = usize::try_from(bucket_to_length(bucket)).unwrap();
     let start = size_class * index;
